+++ src/Codeception/Module/WPWebDriver.php
@@ -25,6 +25,7 @@ requireCodeceptionModules('WPWebDriver', [ 'WebDriver' ]);
 class WPWebDriver extends WebDriver
 {
     use WPBrowserMethods;
+    use ThemeMethods;
 
     /**
      * The module required fields, to be set in the suite .yml configuration file.